requirements.txt
flask==2.3.3
werkzeug==2.3.7
google-generativeai==0.3.0
google-ai-generativelanguage==0.4.0
google-api-core==2.11.0
google-auth==2.22.0
protobuf==4.24.0
tqdm==4.65.0
gunicorn==20.1.0 